reorganize update tests
