+++ backend/src/domain/repositories/recipe/in_memory/tests.rs
@@ -43,26 +43,6 @@ async fn deleting_a_nonexistent_recipe_fails() {
 }
 
 #[tokio::test]
-async fn updating_a_recipe_succeeds() {
-    let ingredient_repo = InMemoryIngredientRepository::new();
-    let repo = InMemoryRecipeRepository::new();
-    __test__::updating_a_recipe_succeeds(repo, ingredient_repo).await
-}
-
-#[tokio::test]
-async fn updating_a_nonexistent_recipe_fails() {
-    let repo = InMemoryRecipeRepository::new();
-    __test__::updating_a_nonexistent_recipe_fails(repo).await
-}
-
-#[tokio::test]
-async fn updating_a_recipe_with_empty_changeset_does_nothing() {
-    let ingredient_repo = InMemoryIngredientRepository::new();
-    let repo = InMemoryRecipeRepository::new();
-    __test__::updating_a_recipe_with_empty_changeset_does_nothing(repo, ingredient_repo).await
-}
-
-#[tokio::test]
 async fn adding_an_ingredient_to_a_recipe_works() {
     let ingredient_repo = InMemoryIngredientRepository::new();
     let repo = InMemoryRecipeRepository::new();
