+++ backend/src/domain/repositories/recipe/postgres/tests.rs
@@ -46,26 +46,6 @@ async fn deleting_a_nonexistent_recipe_fails(pool: PgPool) {
 }
 
 #[sqlx::test]
-async fn updating_a_recipe_succeeds(pool: PgPool) {
-    let ingredient_repo = PostgresIngredientRepository::new(pool.clone());
-    let repo = PostgresRecipeRepository::new(pool);
-    __test__::updating_a_recipe_succeeds(repo, ingredient_repo).await
-}
-
-#[sqlx::test]
-async fn updating_a_nonexistent_recipe_fails(pool: PgPool) {
-    let repo = PostgresRecipeRepository::new(pool);
-    __test__::updating_a_nonexistent_recipe_fails(repo).await
-}
-
-#[sqlx::test]
-async fn updating_a_recipe_with_empty_changeset_does_nothing(pool: PgPool) {
-    let ingredient_repo = PostgresIngredientRepository::new(pool.clone());
-    let repo = PostgresRecipeRepository::new(pool);
-    __test__::updating_a_recipe_with_empty_changeset_does_nothing(repo, ingredient_repo).await
-}
-
-#[sqlx::test]
 async fn adding_an_ingredient_to_a_recipe_works(pool: PgPool) {
     let ingredient_repo = PostgresIngredientRepository::new(pool.clone());
     let repo = PostgresRecipeRepository::new(pool);
